app/requirements.txt
aiofiles==0.4.0
aiosqlite==0.11.0
ciso8601==2.1.2
Click==7.0
h11==0.9.0
PyPika==0.35.19
python-multipart==0.0.5
six==1.14.0
starlette==0.13.0
tortoise-orm==0.15.7
typing-extensions==3.7.4.1
uvicorn==0.11.1
websockets==8.1
